change renderMessage fn to base class
as preparation to create menu items in subclasses
