+++ src/platform/web/ui/session/room/timeline/MissingAttachmentView.js
@@ -14,12 +14,11 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {TemplateView} from "../../../general/TemplateView.js";
-import {renderMessage} from "./common.js";
+import {BaseMessageView} from "./BaseMessageView.js";
 
-export class MissingAttachmentView extends TemplateView {
-    render(t, vm) {
+export class MissingAttachmentView extends BaseMessageView {
+    renderMessageBody(t, vm) {
         const remove = t.button({className: "link", onClick: () => vm.abortSending()}, vm.i18n`Remove`);
-        return renderMessage(t, vm, t.p({className: "Timeline_messageBody statusMessage"}, [vm.label, " ", remove]));
+        return t.p({className: "Timeline_messageBody statusMessage"}, [vm.label, " ", remove]);
     }
 }
